feat: criação da classe TicketService, criação do record DadosDetalharTicket e modificação na classe TicketController

+++ src/main/java/Projeto/Gerenciador_Eventos/dto/DadosCadastroTicket.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.NotNull;
 public record DadosCadastroTicket(
 		
 		@NotNull
-		Integer idInscricao,
+		Long idInscricao,
 		
 		@NotBlank
 		String codigoHashTicket) {
